// src/app/model/eventy.ts
export interface Eventy {

  id: number;              // Unique identifier of the event
  title: string;           // Event title
  description: string;     // Event description
  date: Date;              // Event date
  location: string;            // Location
  price: number;            // Ticket price
  organizerId: number;  // Organizer identifier
  imageUrl: string;        // Path/URL to the event image
  nbPlaces: number;        // Number of available places
  nbrLike: number;
  isBest?: boolean;
}